Optimize IC generation
early_sync on GPU
protect makeGlobalBox against empty local domain
protect nodeCountsKernel against empty grid launches
added DeviceVector:shrink_to_fit

+++ domain/include/cstone/focus/octree_focus_mpi.hpp
@@ -450,7 +450,7 @@ public:
             const KeyType* nodeKeys = treeData_.prefixes.data();
 
 #pragma omp parallel for schedule(static)
-            for (size_t i = 0; i < treeData_.numNodes; ++i)
+            for (int i = 0; i < treeData_.numNodes; ++i)
             {
                 //! set centers to geometric centers for min dist Mac
                 centers_[i] = computeMinMacR2(nodeKeys[i], invThetaEff, box_);
